pyproject.toml
[project]
name = "nacc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.1.1",
    "datasets>=3.1.0",
    "ipdb>=0.13.13",
    "loguru>=0.7.3",
    "nltk>=3.9.1",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "peft>=0.13.2",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "torch>=2.5.1",
    "transformers>=4.46.2",
    "wandb>=0.18.6",
    "huggingface-hub>=0.26.2",
]

